+++ concordium-sdk/src/main/java/com/concordium/sdk/transactions/tokens/TransferTokenOperation.java
@@ -2,29 +2,40 @@ package com.concordium.sdk.transactions.tokens;
 
 import com.concordium.sdk.types.UInt64;
 import lombok.Builder;
+import lombok.EqualsAndHashCode;
 import lombok.Getter;
-import lombok.val;
+import lombok.NonNull;
+import lombok.extern.jackson.Jacksonized;
 
-import java.util.HashMap;
 import java.util.Optional;
 
+/**
+ * Transfers a specified amount of tokens from the sender account to the recipient account.
+ *
+ * @see <a href="https://github.com/Concordium/concordium-update-proposals/blob/main/source/CIS/cis-7.rst#transfer">CBOR Schema</a>
+ */
 @Getter
 @Builder
+@EqualsAndHashCode
+// This is for CBOR serialization to use the generated builder.
+@Jacksonized
 public class TransferTokenOperation implements TokenOperation {
 
     /**
      * Amount to be transferred.
      * It very important that the decimals in it match the actual value of the token.
      */
+    @NonNull
     private final TokenOperationAmount amount;
 
     /**
      * Recipient of the transfer.
      */
+    @NonNull
     private final TaggedTokenHolderAccount recipient;
 
     /**
-     * Optional memo (message) to be included to the transfer,
+     * Optional memo (message, data) to be included to the transfer,
      * which will be <b>publicly available</b> on the blockchain.
      */
     private final CborMemo memo;
@@ -34,21 +45,14 @@ public class TransferTokenOperation implements TokenOperation {
     }
 
     @Override
-    public String getType() {
-        return "transfer";
-    }
-
-    @Override
     public UInt64 getBaseCost() {
         return UInt64.from(100);
     }
 
     @Override
-    public Object getBody() {
-        val body = new HashMap<String, Object>();
-        body.put("amount", amount);
-        body.put("recipient", recipient);
-        body.put("memo", memo);
-        return body;
+    public String getType() {
+        return TYPE;
     }
+
+    public static final String TYPE = "transfer";
 }