+++ concordium-sdk/src/main/java/com/concordium/sdk/types/UInt64.java
@@ -1,11 +1,10 @@
 package com.concordium.sdk.types;
 
-import com.concordium.sdk.serializing.CborMapper;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import com.fasterxml.jackson.databind.ser.std.StdSerializer;
 import com.fasterxml.jackson.dataformat.cbor.CBORGenerator;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -17,7 +16,10 @@ import java.nio.ByteBuffer;
 
 @Getter
 @EqualsAndHashCode
-@JsonSerialize(using = UInt64.UInt64Serializer.class)
+@JsonSerialize(
+        using = UInt64.UInt64Serializer.class
+        // And deserialization works via the @JsonCreator constructor.
+)
 public final class UInt64 implements Comparable<UInt64> {
     public static final int BYTES = Long.BYTES;
 
@@ -80,9 +82,13 @@ public final class UInt64 implements Comparable<UInt64> {
 
     /**
      * A custom Jackson serializer is provided that ensures that the unsigned value
-     * is the one used when serializing to JSON.
+     * is the one used when serializing to JSON and CBOR.
      */
-    static class UInt64Serializer extends JsonSerializer<UInt64> {
+    static class UInt64Serializer extends StdSerializer<UInt64> {
+
+        protected UInt64Serializer() {
+            super(UInt64.class);
+        }
 
         @Override
         public void serialize(UInt64 uint,
@@ -90,15 +96,7 @@ public final class UInt64 implements Comparable<UInt64> {
                               SerializerProvider provider) throws IOException {
 
             if (generator instanceof CBORGenerator) {
-                val cborGenerator = (CBORGenerator) generator;
-
-                // Write as unsigned 8-byte integer.
-                val value = ByteBuffer.allocate(1 + Long.BYTES)
-                        .put((byte) 0x1B)
-                        .put(uint.getBytes())
-                        .array();
-                CborMapper.writeBytesAsValue(cborGenerator, value, 0, value.length);
-
+                ((CBORGenerator) generator).writeNumberUnsigned(uint.value);
                 return;
             }