+++ concordium-sdk/src/test/java/com/concordium/sdk/tokens/TokenOperationAmountTest.java
@@ -4,6 +4,7 @@ import com.concordium.sdk.serializing.CborMapper;
 import com.concordium.sdk.transactions.tokens.TokenOperationAmount;
 import com.concordium.sdk.types.UInt64;
 import lombok.SneakyThrows;
+import lombok.val;
 import org.bouncycastle.util.encoders.Hex;
 import org.junit.Assert;
 import org.junit.Test;
@@ -13,60 +14,111 @@ public class TokenOperationAmountTest {
     @Test
     @SneakyThrows
     public void testTokenOperationAmountSerialization() {
+        val amountA = new TokenOperationAmount(
+                UInt64.from("1500000"),
+                6
+        );
+        val amountAExpectedHex = "c482251a0016e360";
         Assert.assertEquals(
-                "c482251b000000000016e360",
-                Hex.toHexString(
-                        CborMapper.INSTANCE.writeValueAsBytes(
-                                new TokenOperationAmount(
-                                        UInt64.from("1500000"),
-                                        6
-                                )
-                        )
-                )
+                amountAExpectedHex,
+                Hex.toHexString(CborMapper.INSTANCE.writeValueAsBytes(amountA))
         );
         Assert.assertEquals(
-                "c482221b000000000012d687",
-                Hex.toHexString(
-                        CborMapper.INSTANCE.writeValueAsBytes(
-                                new TokenOperationAmount(
-                                        UInt64.from("1234567"),
-                                        3
-                                )
-                        )
+                amountA,
+                CborMapper.INSTANCE.readValue(
+                        Hex.decode(amountAExpectedHex),
+                        TokenOperationAmount.class
                 )
         );
+        val amountB = new TokenOperationAmount(
+                UInt64.from("1234567"),
+                3
+        );
+        val amountBExpectedHex = "c482221a0012d687";
         Assert.assertEquals(
-                "c482221b8000000000000000",
-                Hex.toHexString(
-                        CborMapper.INSTANCE.writeValueAsBytes(
-                                new TokenOperationAmount(
-                                        UInt64.from(Long.MIN_VALUE),
-                                        3
-                                )
-                        )
-                )
+                amountBExpectedHex,
+                Hex.toHexString(CborMapper.INSTANCE.writeValueAsBytes(amountB))
         );
         Assert.assertEquals(
-                "c482221b7fffffffffffffff",
-                Hex.toHexString(
-                        CborMapper.INSTANCE.writeValueAsBytes(
-                                new TokenOperationAmount(
-                                        UInt64.from(Long.MAX_VALUE),
-                                        3
-                                )
-                        )
+                amountB,
+                CborMapper.INSTANCE.readValue(
+                        Hex.decode(amountBExpectedHex),
+                        TokenOperationAmount.class
                 )
         );
+        val amountC =  new TokenOperationAmount(
+                UInt64.from(Long.MIN_VALUE),
+                3
+        );
+        val amountCExpectedHex = "c482221b8000000000000000";
         Assert.assertEquals(
-                "c482221bffffffffffffffff",
-                Hex.toHexString(
-                        CborMapper.INSTANCE.writeValueAsBytes(
-                                new TokenOperationAmount(
-                                        UInt64.from(-1),
-                                        3
-                                )
-                        )
+                amountCExpectedHex,
+                Hex.toHexString(CborMapper.INSTANCE.writeValueAsBytes(amountC))
+        );
+        Assert.assertEquals(
+                amountC,
+                CborMapper.INSTANCE.readValue(
+                        Hex.decode(amountCExpectedHex),
+                        TokenOperationAmount.class
                 )
+        );
+        val amountD =  new TokenOperationAmount(
+                UInt64.from(Long.MAX_VALUE),
+                3
+        );
+        val amountDExpectedHex = "c482221b7fffffffffffffff";
+        Assert.assertEquals(
+                amountDExpectedHex,
+                Hex.toHexString(CborMapper.INSTANCE.writeValueAsBytes(amountD))
+        );
+        Assert.assertEquals(
+                amountD,
+                CborMapper.INSTANCE.readValue(
+                        Hex.decode(amountDExpectedHex),
+                        TokenOperationAmount.class
+                )
+        );
+        val amountE = new TokenOperationAmount(
+                UInt64.from(-1),
+                3
+        );
+        val amountEExpectedHex = "c482221bffffffffffffffff";
+        Assert.assertEquals(
+                amountEExpectedHex,
+                Hex.toHexString(CborMapper.INSTANCE.writeValueAsBytes(amountE))
+        );
+        Assert.assertEquals(
+                amountE,
+                CborMapper.INSTANCE.readValue(
+                        Hex.decode(amountEExpectedHex),
+                        TokenOperationAmount.class
+                )
+        );
+    }
+
+    @Test
+    @SneakyThrows
+    public void testTokenOperationAmountListSerialization() {
+        val amounts = new TokenOperationAmount[]{
+                new TokenOperationAmount(
+                        UInt64.from("1500000"),
+                        6
+                ),
+                new TokenOperationAmount(
+                        UInt64.from("1234567"),
+                        3
+                )
+        };
+        val expectedHex = "82c482251a0016e360c482221a0012d687";
+        Assert.assertEquals(
+                expectedHex,
+                Hex.toHexString(CborMapper.INSTANCE.writeValueAsBytes(amounts))
+        );
+        Assert.assertArrayEquals(
+                amounts,
+                CborMapper.INSTANCE
+                        .readerForArrayOf(TokenOperationAmount.class)
+                        .readValue(Hex.decode(expectedHex))
         );
     }
 }