+++ concordium-sdk/src/test/java/com/concordium/sdk/transactions/TokenUpdateTransactionTest.java
@@ -1,6 +1,5 @@
 package com.concordium.sdk.transactions;
 
-import com.concordium.grpc.v2.plt.TokenId;
 import com.concordium.sdk.transactions.tokens.CborMemo;
 import com.concordium.sdk.transactions.tokens.TaggedTokenHolderAccount;
 import com.concordium.sdk.transactions.tokens.TokenOperationAmount;
@@ -9,6 +8,8 @@ import com.concordium.sdk.types.AccountAddress;
 import com.concordium.sdk.types.Nonce;
 import com.concordium.sdk.types.UInt64;
 import lombok.SneakyThrows;
+import lombok.val;
+import org.bouncycastle.util.encoders.Hex;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -20,7 +21,7 @@ public class TokenUpdateTransactionTest {
     @SneakyThrows
     public void testTokenUpdateTransferTransactionWithTextMemo() {
         Assert.assertEquals(
-                "24303123d37c83a071bf861bb5d4490f945e4253234890062d851fd75c4bf25c",
+                "9f743fdb00d9697cde98dc3e75621b40535ec84b953e83f8a4116cb2c61519c0",
                 TransactionFactory
                         .newTokenUpdate()
                         .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
@@ -62,7 +63,7 @@ public class TokenUpdateTransactionTest {
     @SneakyThrows
     public void testTokenUpdateTransferTransactionWithoutMemo() {
         Assert.assertEquals(
-                "76d7e2c3f3e2046eb145783f7bda8c26c216e2f2db3a195f41ffb7640981077a",
+                "3945fe655328c62cc75be7e06cc24c91336b0bd4854d930ed017757c42a3ffc2",
                 TransactionFactory
                         .newTokenUpdate()
                         .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
@@ -101,7 +102,7 @@ public class TokenUpdateTransactionTest {
     @SneakyThrows
     public void testTokenUpdateTransferTransactionCostWithoutMemo() {
         Assert.assertEquals(
-                UInt64.from("751"),
+                UInt64.from("745"),
                 TransactionFactory
                         .newTokenUpdate()
                         .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
@@ -133,6 +134,77 @@ public class TokenUpdateTransactionTest {
                         .build()
                         .getHeader()
                         .getMaxEnergyCost()
+        );
+    }
+
+    @Test
+    @SneakyThrows
+    public void testTokenUpdateTransferTransactionCostWithMemo() {
+        Assert.assertEquals(
+                UInt64.from("828"),
+                TransactionFactory
+                        .newTokenUpdate()
+                        .sender(AccountAddress.from("3JwD2Wm3nMbsowCwb1iGEpnt47UQgdrtnq2qT6opJc3z2AgCrc"))
+                        .payload(
+                                TokenUpdate
+                                        .builder()
+                                        .tokenSymbol("TEST")
+                                        .operation(
+                                                TransferTokenOperation
+                                                        .builder()
+                                                        .amount(
+                                                                new TokenOperationAmount(
+                                                                        new BigDecimal("1.5"),
+                                                                        2
+                                                                )
+                                                        )
+                                                        .recipient(
+                                                                new TaggedTokenHolderAccount(
+                                                                        AccountAddress.from("3hYXYEPuGyhFcVRhSk2cVgKBhzVcAryjPskYk4SecpwGnoHhuM")
+                                                                )
+                                                        )
+                                                        .memo(
+                                                                CborMemo.from("Memo for the test, long memo, quite a long text for a memo if you ask me")
+                                                        )
+                                                        .build()
+                                        )
+                                        .build()
+                        )
+                        .nonce(Nonce.from(78910))
+                        .expiry(Expiry.from(123456))
+                        .signer(TransactionTestHelper.getValidSigner())
+                        .build()
+                        .getHeader()
+                        .getMaxEnergyCost()
+        );
+    }
+
+    @Test
+    @SneakyThrows
+    public void testTokenUpdateBuilderWithSerializedOperations() {
+        val expectedPayload = TokenUpdate
+                .builder()
+                .tokenSymbol("TEST")
+                .operation(
+                        TransferTokenOperation
+                                .builder()
+                                .amount(new TokenOperationAmount(new BigDecimal("1.5"), 6))
+                                .recipient(new TaggedTokenHolderAccount(
+                                        AccountAddress.from(
+                                                "3CbvrNVpcHpL7tyT2mhXxQwNWHiPNYEJRgp3CMgEcMyXivms6B"
+                                        )
+                                ))
+                                .build()
+                )
+                .build();
+        val serializedOperationsHex = Hex.toHexString(expectedPayload.getOperationsSerialized());
+        Assert.assertEquals(
+                expectedPayload,
+                TokenUpdate
+                        .builder()
+                        .tokenSymbol("TEST")
+                        .operationsSerialized(Hex.decode(serializedOperationsHex))
+                        .build()
         );
     }
 }